+++ src/string.c
@@ -28,7 +28,7 @@ memcmp(const void *v1, const void *v2, uint n)
 }
 
 void*
-memmove(void *dst, const void *src, uint n)
+memmove(void *dst, const void *src, size_t n)
 {
   const char *s;
   char *d;
